[step2] refactor: score 계산로직을 Cards에 위임
