+++ src/main/kotlin/blackjack/view/output/PlayerResultOutputView.kt
@@ -7,7 +7,7 @@ class PlayerResultOutputView(player: Player) {
         with(player) {
             val message = "${name.value}카드: "
             val cardsStr = cards.value.map { card -> card.denom.symbol + card.type.korName }.joinToString(SEPARATOR)
-            println(message + cardsStr + " - 결과: ${player.getOptimizedScore()}")
+            println(message + cardsStr + " - 결과: ${player.cards.getOptimizedScore()}")
         }
     }
